Update WooWearComplicationService.kt to only be responsible for Orders Count
