+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/complications/OrderCountComplicationService.kt
@@ -1,0 +1,36 @@
+package com.woocommerce.android.wear.complications
+
+import android.util.Log
+import androidx.wear.watchface.complications.data.ComplicationData
+import androidx.wear.watchface.complications.data.ComplicationType
+import androidx.wear.watchface.complications.datasource.ComplicationRequest
+import androidx.wear.watchface.complications.datasource.SuspendingComplicationDataSourceService
+import com.woocommerce.android.R
+
+class OrderCountComplicationService : SuspendingComplicationDataSourceService() {
+    override fun getPreviewData(type: ComplicationType): ComplicationData? {
+        return when (type) {
+            ComplicationType.SHORT_TEXT -> {
+                createTextComplicationData(
+                    context = applicationContext,
+                    content = "5",
+                    description = getString(R.string.orders_complication_preview_description)
+                )
+            }
+
+            else -> null
+        }
+    }
+
+    override suspend fun onComplicationRequest(request: ComplicationRequest): ComplicationData? {
+        return when (request.complicationType) {
+            ComplicationType.SHORT_TEXT -> createTextComplicationData(
+                context = applicationContext,
+                content = "10", // get Today's Orders count
+                description = getString(R.string.orders_complication_preview_description)
+            )
+
+            else -> null
+        }
+    }
+}
